requirements.txt
flask==2.2.2
Werkzeug==2.3.7
line-bot-sdk
openai
requests
pyimgure